requirements.txt
streamlit
pandas
plotly.express
random
logging
time
